Update routing to run from _layout and delete index file
